project 3:  TV script generation files
Please note that on this project, I spent quite a bit of time in the theory section but then was really lost in the high level abstraction of tensorflow implementation so spent a lot of time in forums and actually based my code on what was there.  I really need to spend more time on reading documentation for tensorflow and numpy (I started late so skipped the numpy section at the beginning of the course) but feel I have enough foundational skills in linear algebra and the theories (also bought and reading Andrew Trask's book) to follow along.  Also, I need to read Python documentation and more practice to use things like "with" statements in my code.  My background is in C, C++ and Java so I understand most of Python but sytanxically, I am not yet that proficient.  Having said that, I am a translator by trade so will be trying to build my own RNN based on course materials as a personal project I can use in my work and to show prospective employers.